ci: fix Windows test build and lint by lowering Go directive
Two CI failures observed on the first run:

1. Windows test job failed `go vet` because `syscall.Kill` and
   `syscall.SIGTERM` are not defined on Windows. The two tests that
   used them (`TestStartServerDaemonHandlesConcurrentClients` and
   `TestStartServerGracefulShutdown`) now live in a new
   `net_unix_test.go` gated with `//go:build unix`. Windows still
   runs `TestStartServerProxiesSingleConnection` from `net_test.go`,
   plus every other unit and integration test.

2. Lint job failed because golangci-lint v1.64.8 (built with Go
   1.24) refused to load a module that targeted Go 1.26.3. Lowered
   the go directive to 1.23 — well within reach of the toolchain
   the linter ships with. Indirect deps downgraded to versions that
   still target Go 1.23 (golang.org/x/crypto v0.31.0,
   golang.org/x/sys v0.28.0). All tests still pass with -race and
   cross-vet for GOOS=windows is clean.
