+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/gedigi/noisecat
 
-go 1.26.3
+go 1.23
 
 require (
 	github.com/flynn/noise v1.1.0
@@ -8,6 +8,6 @@ require (
 )
 
 require (
-	golang.org/x/crypto v0.52.0 // indirect
-	golang.org/x/sys v0.45.0 // indirect
+	golang.org/x/crypto v0.31.0 // indirect
+	golang.org/x/sys v0.28.0 // indirect
 )
